third_party/blink/web_tests/external/wpt/editing/run/insertparagraph_4001-5000-expected.txt

This is a testharness.js-based test.
Found 15 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<p>foo<span style=color:#aBcDeF>[bar]</span>baz" queryCommandState("stylewithcss") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<p>foo<span style=color:#aBcDeF>[bar]</span>baz" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertparagraph",""]] "<p>foo<span style=color:#aBcDeF>[bar]</span>baz" queryCommandValue("defaultparagraphseparator") after
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<p>foo<span style=color:#aBcDeF>{bar}</span>baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertparagraph",""]] "<p>foo<span style=color:#aBcDeF>{bar}</span>baz" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["insertparagraph",""]] "<ul contenteditable><li>{}<br></ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul contenteditable=\\"\\"><li><br></li></ul>" but got "<p><br></p>"
[FAIL] [["insertparagraph",""]] "<div contenteditable=false><ul contenteditable><li>{}<br></ul></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div contenteditable=\\"false\\"><ul contenteditable=\\"\\"><li><br></li></ul></div>" but got "<div contenteditable=\\"false\\"><ul contenteditable=\\"\\"><li><br></li><li><br></li></ul></div>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<dl><dt><p>foo[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt><p>foo</p></dt><dd><p><br></p></dd></dl>" but got "<dl><dt><p>foo</p></dt><dt><p><br></p></dt></dl>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<dl><dt><p>foo[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt><p>foo</p></dt><dd><p><br></p></dd></dl>" but got "<dl><dt><p>foo</p></dt><dt><p><br></p></dt></dl>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<dl><dd><p>foo[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd><p>foo</p></dd><dt><p><br></p></dt></dl>" but got "<dl><dd><p>foo</p></dd><dd><p><br></p></dd></dl>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<dl><dd><p>foo[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd><p>foo</p></dd><dt><p><br></p></dt></dl>" but got "<dl><dd><p>foo</p></dd><dd><p><br></p></dd></dl>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<dl><dt><div>foo[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt><div>foo</div></dt><dd><div><br></div></dd></dl>" but got "<dl><dt><div>foo</div></dt><dt><div><br></div></dt></dl>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<dl><dt><div>foo[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt><div>foo</div></dt><dd><div><br></div></dd></dl>" but got "<dl><dt><div>foo</div></dt><dt><div><br></div></dt></dl>"
[FAIL] [["defaultparagraphseparator","div"],["insertparagraph",""]] "<dl><dd><div>foo[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd><div>foo</div></dd><dt><div><br></div></dt></dl>" but got "<dl><dd><div>foo</div></dd><dd><div><br></div></dd></dl>"
[FAIL] [["defaultparagraphseparator","p"],["insertparagraph",""]] "<dl><dd><div>foo[]</dl>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd><div>foo</div></dd><dt><div><br></div></dt></dl>" but got "<dl><dd><div>foo</div></dd><dd><div><br></div></dd></dl>"
Harness: the test ran to completion.
